Re #17: Separate self-reply loading from others
By splitting "to:user from:user" from "to:user from:<others>" we can
ensure that a popular thread that the author replied to later
doesn't get overrun by the responses.

Still need to improve the UI, and we're still doomed for things like
GitHub where they get SO many mentions that we'll rarely
get a decent set of replies from other people.