Re #17: Handle adding different duplicate types
If a tweet exists as priority and gets re-added as not then we
ignore it. If it exists as non-priority and gets re-added as priority
then we move it. We already handled being added as the same
type by ignoring it.

To do this, we've also got a refactored "contains_id"
and a new "index_of" function